backend/requirements.txt
# FastAPI and server
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.1
python-multipart>=0.0.5

# MongoDB
motor>=2.5.1
pymongo>=3.12.0

# OpenAI
openai>=1.0.0

# Computer Vision
mediapipe>=0.8.9
opencv-python>=4.5.3
numpy>=1.21.2
matplotlib>=3.4.3

# Optional - for testing
pytest>=6.2.5
httpx>=0.19.0